Feat: sad path for market or vendor not being valid for market_vendor

+++ app/controllers/api/v0/market_vendors_controller.rb
@@ -26,10 +26,19 @@ class Api::V0::MarketVendorsController < ApplicationController
   def destroy
     market = Market.find_by_id(params[:market_vendor][:market_id])
     vendor = Vendor.find_by_id(params[:market_vendor][:vendor_id])
-    
-    market_vendor = MarketVendor.find_market_vendor(market, vendor)
 
-    market_vendor.destroy_all
+    if market.nil? || vendor.nil?
+      render json: {
+        "errors": [
+          { 
+            "detail": "No MarketVendor with market_id=#{params[:market_vendor][:market_id]} AND vendor_id=#{params[:market_vendor][:vendor_id]} exists"
+            }]
+          },
+      status: 404
+    else
+      market_vendor = MarketVendor.find_market_vendor(market, vendor)
+      market_vendor.destroy_all
+    end
   end
 
   private
